# Extract values on the fly