--- a/src/token.jl
+++ b/src/token.jl
@@ -42,7 +42,9 @@ TOKEN_ERROR_DESCRIPTION = Dict{TokenError, String}(
     UNKNOWN => "unknown",
 )
 
-struct Token
+abstract type AbstractToken end
+
+struct Token <: AbstractToken
     kind::Kind
     # Offsets into a string or buffer
     startpos::Tuple{Int, Int} # row, col where token starts /end, col is a string index
@@ -53,6 +55,15 @@ struct Token
     token_error::TokenError
 end
 
+struct RawToken <: AbstractToken
+    kind::Kind
+    # Offsets into a string or buffer
+    startpos::Tuple{Int, Int} # row, col where token starts /end, col is a string index
+    endpos::Tuple{Int, Int}
+    startbyte::Int # The byte where the token start in the buffer
+    endbyte::Int # The byte where the token ended in the buffer
+end
+
 function Token(kind::Kind, startposition::Tuple{Int, Int}, endposition::Tuple{Int, Int},
                startbyte::Int, endbyte::Int, val::String)
     Token(kind, startposition, endposition, startbyte, endbyte, val, NO_ERR)
@@ -61,16 +72,16 @@ Token() = Token(ERROR, (0,0), (0,0), 0, 0, "", UNKNOWN)
 
 const EMPTY_TOKEN = Token()
 
-function kind(t::Token)
+function kind(t::AbstractToken)
     isoperator(t.kind) && return OP
     iskeyword(t.kind) && return KEYWORD
     return t.kind
 end
-exactkind(t::Token) = t.kind
-startpos(t::Token) = t.startpos
-endpos(t::Token) = t.endpos
+exactkind(t::AbstractToken) = t.kind
+startpos(t::AbstractToken) = t.startpos
+endpos(t::AbstractToken) = t.endpos
 function untokenize(t::Token)
-    if t.kind == IDENTIFIER || isliteral(t.kind)
+    if t.kind == IDENTIFIER || isliteral(t.kind) || t.kind == COMMENT || t.kind == WHITESPACE || t.kind == ERROR
         return t.val
     elseif iskeyword(t.kind)
         return lowercase(string(t.kind))
@@ -99,8 +110,12 @@ function untokenize(t::Token)
     end
 end
 
+function untokenize(t::RawToken, str::String)
+    String(str[1 + (t.startbyte:t.endbyte)])
+end
+
 function untokenize(ts)
-    if eltype(ts) != Token
+    if eltype(ts) != AbstractToken
         throw(ArgumentError("element type of iterator has to be Token"))
     end
     io = IOBuffer()
@@ -121,5 +136,12 @@ function Base.show(io::IO, t::Token)
 end
 
 Base.print(io::IO, t::Token) = print(io, untokenize(t))
+
+function Base.show(io::IO, t::RawToken)
+    start_r, start_c = startpos(t)
+    end_r, end_c = endpos(t)
+    print(io, rpad(string(start_r, ",", start_c, "-", end_r, ",", end_c), 17, " "))
+    print(io, rpad(kind(t), 15, " "))
+end
 
 end # module