--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -3,4 +3,6 @@ using Base.Test
 import Tokenize
 
 include("lex_yourself.jl")
+@testset "lexer" begin
 include("lexer.jl")
+end