Here's the pseudo val-less branch I've been using, only extracting strings for identifiers and literals.

Some timings:
```
using Tokenize, BenchmarkTools
function speed_test()
    cnt = 0
    dir = dirname(Base.find_source_file("base.jl"))
    for (root, dirs, files) in walkdir(dir)
        for file in files
            if endswith(file, ".jl")
                file = joinpath(root, file)
                # println(file)
                str = readstring(file)
                l = tokenize(str)
                while !Tokenize.Lexers.eof(l)
                    t = Tokenize.Lexers.next_token(l)
                    if t.kind == Tokens.IDENTIFIER
                        cnt += 1
                    end
                end
            end
        end
    end
    cnt
end
@benchmark speed_test()
```
Before:
```
BenchmarkTools.Trial:
  memory estimate:  254.06 MiB
  allocs estimate:  5078640
  --------------
  minimum time:     615.183 ms (5.96% GC)
  median time:      618.250 ms (6.50% GC)
  mean time:        620.669 ms (6.29% GC)
  maximum time:     637.570 ms (6.46% GC)
  --------------
  samples:          9
  evals/sample:     1
```

After:
```
BenchmarkTools.Trial:
  memory estimate:  154.10 MiB
  allocs estimate:  2633640
  --------------
  minimum time:     333.714 ms (6.79% GC)
  median time:      338.120 ms (6.88% GC)
  mean time:        337.438 ms (6.87% GC)
  maximum time:     340.331 ms (6.85% GC)
  --------------
  samples:          15
  evals/sample:     1
```

and obviously it disables the round-tripping tests