@KristofferC 2017-08-30
Nice. This is a little bit too tailor made for the parser though? I would prefer to have two modes:

1) Completely roundtrippable (but still as efficient as possible). This can be done similar to how it is done here, except also store `val` for whitespace and comments.

2) No vals at all. Tokens can then be stack allocated, inlined into vectors and there is no need to write to any buffer. If you need a `val` later, you can extract it from the original string by using the byte offsets in each Token.

Opinions?