@ZacLN 2017-08-30
Yep that sounds fine w/ me, we skip ws/comments internally in CSTParser anyway. Do we need to add  `RawToken <: Token`  and `ValToken <: Token`for the second case?