@KristofferC 2017-08-30
For backwards compatability, I think we should keep `Token` as is, add an `AbstractToken` and a `RawToken`.