@Keno 2017-08-30
> I can just parameterize Lexer if you like

I think that would be the best way to go about it.