@ZacLN 2017-08-30
Ta, fixes made